Add VK group fix

+++ tg-bot/src/main/java/polis/commands/TgSyncGroups.java
@@ -94,7 +94,7 @@ public class TgSyncGroups extends Command {
                                                     currentAccount.getAccessToken(),
                                                     (int) currentAccount.getAccountId()
                                             ),
-                                            group.getGroupId()
+                                            (int) group.getGroupId()
                                     );
                                     break;
                                 }
